feat: Add entity inspector with property panel and event handling
- Implemented a property panel example for displaying entity components.
- Created event system for entity inspector updates, including entity and component events.
- Developed reflection utilities for extracting and formatting component data.
- Built tree structures for the inspector UI, grouping components by crate.
- Designed a property panel UI for displaying detailed component information.
- Established UI management systems for handling inspector events and tree updates.
- Added camera setup for inspector rendering.
- Implemented tree node selection handling to update the property panel.
