+++ crates/bevy_entity_inspector/src/ui/mod.rs
@@ -4,10 +4,13 @@ use bevy::prelude::*;
 
 /// Disclosure triangle UI components
 pub mod disclosure;
+/// Property panel for detailed component inspection
+pub mod property_panel;
 /// Tree view UI components  
 pub mod tree;
 
 pub use disclosure::*;
+pub use property_panel::*;
 pub use tree::*;
 
 /// Plugin that provides all inspector UI widgets
